refact: Member 도메인으로 매핑 엔티티 이동
