+++ src/main/java/umc/cockple/demo/domain/member/domain/Member.java
@@ -10,8 +10,6 @@ import umc.cockple.demo.domain.notification.domain.Notification;
 import umc.cockple.demo.enums.Gender;
 import umc.cockple.demo.enums.Level;
 import umc.cockple.demo.enums.MemberStatus;
-import umc.cockple.demo.mapping.MemberExercise;
-import umc.cockple.demo.mapping.MemberParty;
 import umc.cockple.demo.global.common.BaseEntity;
 
 import java.time.LocalDate;
